refactor(extension): share panel URL policies
